Reorganize, pass norm
