+++ src/mand/exec/exec_utils.c
@@ -6,7 +6,7 @@
 /*   By: ogonzale <ogonzale@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/17 18:10:22 by ogonzale          #+#    #+#             */
-/*   Updated: 2023/01/04 20:48:04 by ogonzale         ###   ########.fr       */
+/*   Updated: 2023/01/05 18:58:50 by ogonzale         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -76,154 +76,28 @@ char	**get_envp(t_list *environ)
 	return (envp);
 }
 
-int	line_is_limitor(char *line, char *limitor)
+int
+	handle_child_exit(int exit_status, int last_pipe_exit, int last_pipe)
 {
-	int	length_line;
-	int	length_limitor;
-
-	length_line = ft_strlen(line) - 1;
-	length_limitor = ft_strlen(limitor);
-	if (length_line >= length_limitor
-		&& ft_strncmp(line, limitor, length_line) == 0)
-		return (1);
-	else if (length_line < length_limitor
-		&& ft_strncmp(line, limitor, length_limitor) == 0)
-		return (1);
-	return (0);
-}
-
-int	heredoc_continues(char *limitor)
-{
-	char	*line;
-	int		tmp_fd;
-
-	write(STDOUT_FILENO, "> ", 2);
-	line = get_next_line(STDIN_FILENO);
-	if (line == NULL)
+	if (WIFEXITED(exit_status) && last_pipe == 1)
+		return (WEXITSTATUS(exit_status));
+	if (WIFSIGNALED(exit_status))
 	{
-		printf("msh: warning: here-document delimited by end-of-file (wanted `%s')\n", limitor);
-		return (0);
-	}
-	if (line_is_limitor(line, limitor))
-		return (0);
-	tmp_fd = open(TMP_FILE_HEREDOC, O_WRONLY | O_APPEND | O_CREAT, 0600);
-	if (tmp_fd < 0)
-		terminate(ERR_OPEN, 1);
-	write(tmp_fd, line, ft_strlen(line));
-	if (close(tmp_fd) != 0)
-		terminate(ERR_CLOSE, 1);
-	free(line);
-	line = NULL;
-	return (1);
-}
-
-// BUG: When there is more than one heredoc input, cat << EOF << CAT,
-// weird stuff happens
-int	do_here_doc(int *fd_in, char *limitor, int *did_redirection)
-{
-	while (heredoc_continues(limitor))
-		;
-	if (*did_redirection == 1 && close(*fd_in) != 0)
-		terminate(ERR_CLOSE, 1);
-	*fd_in = open(TMP_FILE_HEREDOC, O_RDONLY | O_CREAT);
-	if (*fd_in < 0)
-	{
-		printf("msh: %s: Error reading file or directory\n", limitor);
-		return (1);
-	}
-	unlink(TMP_FILE_HEREDOC);
-	*did_redirection = 1;
-	return (0);
-}
-
-int	handle_open_file(int *fd_in, int *did_redirection, char *filename)
-{
-	if (*did_redirection == 1 && close(*fd_in) != 0)
-		terminate(ERR_CLOSE, 1);
-	*fd_in = open(filename, O_RDONLY);
-	if (*fd_in < 0)
-	{
-		printf("msh: %s: Error reading file or directory\n", filename);
-		return (1);
-	}
-	*did_redirection = 1;
-	return (0);
-}
-
-int	dup_to_in(int *tmp_fd_in, t_list *command)
-{
-	t_list			*token;
-	t_token_content	*token_content;
-	int				fd_in;
-	int				did_redirection;
-
-	token = ((t_cmd_line_content *)command->content)->word;
-	did_redirection = 0;
-	while (token)
-	{
-		token_content = token->content;
-		if (token_content->type == OPEN_FILE
-			&& handle_open_file(&fd_in, &did_redirection,
-				token_content->word) == 1)
-			return (1);
-		else if (token_content->type == LIMITOR
-			&& do_here_doc(&fd_in, token_content->word,
-				&did_redirection) == 1)
-			return (1);
-		token = token->next;
-	}
-	if (did_redirection == 1 && dup2(fd_in, *tmp_fd_in) == -1)
-		terminate(ERR_DUP, 1);
-	if (did_redirection == 1 && close(fd_in) != 0)
-		terminate(ERR_CLOSE, 1);
-	return (0);
-}
-
-int	handle_write_to_exit_file(int *fd_out, int *did_redirection,
-								t_token_content *token_content)
-{
-	if (token_content->type == EXIT_FILE
-		|| token_content->type == EXIT_FILE_APP)
-	{
-		if (*did_redirection == 1 && close(*fd_out) != 0)
-			terminate(ERR_CLOSE, 1);
-		if (token_content->type == EXIT_FILE)
-			*fd_out = open(token_content->word,
-					O_WRONLY | O_TRUNC | O_CREAT, 0644);
-		else if (token_content->type == EXIT_FILE_APP)
-			*fd_out = open(token_content->word,
-					O_WRONLY | O_APPEND | O_CREAT, 0644);
-		if (*fd_out < 0)
+		if (WCOREDUMP(exit_status))
+			terminate(ERR_CHLD, 0);
+		if (WTERMSIG(exit_status) == SIGINT)
 		{
-			printf("msh: %s: Error writing file or directory\n", token_content->word);
-			return (1);
+			write(STDOUT_FILENO, "\n", 1);
+			if (last_pipe == 1 || (last_pipe == 0 && last_pipe_exit != 0))
+				return (130);
 		}
-		*did_redirection = 1;
+		if (WTERMSIG(exit_status) == SIGQUIT)
+		{
+			write(STDOUT_FILENO, "Quit\n", 5);
+			rl_on_new_line();
+			if (last_pipe == 1 || (last_pipe == 0 && last_pipe_exit != 0))
+				return (131);
+		}
 	}
-	return (0);
-}
-
-int	dup_to_out(int *tmp_fd_out, t_list *command, int *did_out_redirection)
-{
-	t_list			*token;
-	t_token_content	*token_content;
-	int				did_redirection;
-	int				fd_out;
-
-	did_redirection = 0;
-	token = ((t_cmd_line_content *)command->content)->word;
-	while (token)
-	{
-		token_content = token->content;
-		if (handle_write_to_exit_file(&fd_out,
-				&did_redirection, token_content) == 1)
-			return (1);
-		token = token->next;
-	}
-	if (did_redirection == 1 && dup2(fd_out, *tmp_fd_out) == -1)
-		terminate(ERR_DUP, 1);
-	if (did_redirection == 1 && close(fd_out) != 0)
-		terminate(ERR_CLOSE, 1);
-	*did_out_redirection = did_redirection;
-	return (0);
+	return (last_pipe_exit);
 }
